- Fixed bug with the loader and multiple tabs. - Added notification while issue is being reported. - Once an image has been compressed it will no longer try again.

+++ resilience-app/src/main/java/au/com/dius/resilience/util/ImageCompressor.java
@@ -21,6 +21,11 @@ public class ImageCompressor {
     File compressedFile;
     try {
       compressedFile = new File(path.replace(JPG, COMPRESSED_JPG));
+
+      if (compressedFile.exists()) {
+        return compressedFile;
+      }
+
       FileOutputStream fileOutputStream = new FileOutputStream(compressedFile);
 
       BitmapFactory.Options options = optimiseDecoderOptions(path);
